+++ lib/model_editors.dart
@@ -1,11 +1,11 @@
-export 'src/controllers/abstract_list.dart'             show AbstractListEditingController;
-export 'src/controllers/checkbox_group.dart'            show CheckboxGroupEditingController;
-export 'src/controllers/with_id_title.dart'             show WithIdTitleEditingController;
-export 'src/controllers/with_text.dart'                 show WithTextEditingController;
+export 'src/controllers/abstract_list.dart';
+export 'src/controllers/checkbox_group.dart';
+export 'src/controllers/with_id_title.dart';
+export 'src/controllers/with_text.dart';
 
-export 'src/widgets/capsule_list_editor.dart'           show CapsuleListEditor;
-export 'src/widgets/column_list_editor.dart'            show ColumnListEditor;
-export 'src/widgets/list_add_button_builder.dart'       show ListAddButtonBuilder;
-export 'src/widgets/material_checkbox_column.dart'      show MaterialCheckboxColumn;
-export 'src/widgets/reorderable_list_view_editor.dart'  show ReorderableListViewEditor;
-export 'src/widgets/wrap_list_editor.dart'              show WrapListEditor;
+export 'src/widgets/capsule_list_editor.dart';
+export 'src/widgets/column_list_editor.dart';
+export 'src/widgets/list_add_button_builder.dart';
+export 'src/widgets/material_checkbox_column.dart';
+export 'src/widgets/reorderable_list_view_editor.dart';
+export 'src/widgets/wrap_list_editor.dart';